+++ DelvUI/changelog.md
@@ -1,3 +1,6 @@
+# 2.2.0.5
+- Fix Bard's Radiant Finale duration in Party Cooldowns
+
 # 2.2.0.4
 - Added duration labels for Pictomancer's Hammer Time and Hyperphantasia Bars.
 - Fixed Sage's Eukrasian Dyskrasia not being tracked in the Eukrasian Dosis Bar.